Refractor almost the entire project
