+++ Glacier/include/Shader.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "common.hpp"
+#include "Buffer.hpp"
 
 #include <string_view>
 
@@ -8,15 +9,19 @@ namespace glacier
 {
 	class Application;
 
+	/**
+	 * @brief Types of shaders
+	*/
 	enum class ShaderType
 	{
-		Vertex, Fragment
+		Vertex, Tesselation, Geometry, Fragment, Compute
 	};
 
 	class Shader
 	{
 	public:
-		GLACIER_API Shader(const Application& application, std::string_view path, const ShaderType& type);
+		GLACIER_API Shader(const Application* application, std::string_view path);
+		GLACIER_API Shader(const Application* application, const Buffer& buffer);
 		GLACIER_API ~Shader();
 
 		// Delete copy
@@ -24,15 +29,14 @@ namespace glacier
 		Shader& operator=(const Shader&) = delete;
 
 		// Delete move
-		Shader(Shader&&) = delete;
-		Shader& operator=(Shader&&) = delete;
-
-		inline ShaderType getType() const { return m_Type; }
+		Shader(Shader&& other) = delete;
+		Shader& operator=(Shader&& other) = delete;
 	private:
-		const Application& m_Application;
+		const Application* m_Application;
 		void* m_ShaderModule;
-		ShaderType m_Type;
 
 		friend class Application;
+		friend class Renderer;
+		friend class Pipeline;
 	};
 }
