+++ Glacier/include/glacier.hpp
@@ -3,7 +3,8 @@
 #include "Application.hpp"
 #include "Buffer.hpp"
 #include "File.hpp"
-#include "PipelineInfo.hpp"
+#include "Pipeline.hpp"
 #include "Shader.hpp"
 #include "VertexBuffer.hpp"
 #include "Window.hpp"
+#include "Renderer.hpp"
