+++ Glacier/src/Application.cpp
@@ -3,6 +3,8 @@
 
 #include "Application.hpp"
 #include "VertexBuffer.hpp"
+#include "Renderer.hpp"
+#include "internal/utility.hpp"
 
 #include <vector>
 #include <iostream>
@@ -12,44 +14,6 @@
 #include <GLFW/glfw3.h>
 
 constexpr unsigned int MAX_BUFFERED_FRAMES = 2;
-
-struct QueueFamilyIndices
-{
-	std::optional<unsigned int> graphicsFamily;
-	std::optional<unsigned int> presentationFamily;
-
-	bool isComplete()
-	{
-		if (graphicsFamily.has_value() && presentationFamily.has_value())
-			return true;
-
-		return false;
-	}
-};
-
-struct SwapchainSupportDetails
-{
-	VkSurfaceCapabilitiesKHR capabilities;
-	std::vector<VkSurfaceFormatKHR> formats;
-	std::vector<VkPresentModeKHR> presentModes;
-
-	bool isAdequate()
-	{
-		return (formats.empty() || presentModes.empty()) == false;
-	}
-};
-
-struct ShaderInfo
-{
-	VkShaderStageFlagBits stage;
-	VkShaderModule module;
-};
-
-struct BufferInfo
-{
-	VkVertexInputBindingDescription inputBinding;
-	std::vector<VkVertexInputAttributeDescription> inputAttributes;
-};
 
 static VKAPI_ATTR VkBool32 VKAPI_CALL debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT severity, VkDebugUtilsMessageTypeFlagsEXT type, const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData, void* pUserData)
 {
@@ -98,117 +62,6 @@ void vkDestroyDebugUtilsMessengerEXT(VkInstance instance, VkDebugUtilsMessengerE
 		return function(instance, debugMessenger, pAllocator);
 }
 
-QueueFamilyIndices findQueueFamilies(VkPhysicalDevice device, VkSurfaceKHR surface)
-{
-	QueueFamilyIndices queueFamilyIndices;
-
-	unsigned int queueFamilyCount;
-	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
-
-	std::vector<VkQueueFamilyProperties> queueFamilyProperties(queueFamilyCount);
-	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilyProperties.data());
-
-	unsigned int i = 0;
-	for (const VkQueueFamilyProperties& properties : queueFamilyProperties)
-	{
-		if (properties.queueFlags & VK_QUEUE_GRAPHICS_BIT)
-		{
-			queueFamilyIndices.graphicsFamily = i;
-		}
-
-		VkBool32 presentSupport = VK_FALSE;
-		vkGetPhysicalDeviceSurfaceSupportKHR(device, i, surface, &presentSupport);
-
-		if (presentSupport)
-		{
-			queueFamilyIndices.presentationFamily = i;
-		}
-
-		// If all the queue families are found, break the loop
-		if (queueFamilyIndices.isComplete())
-			break;
-
-		i++;
-	}
-
-	return queueFamilyIndices;
-}
-
-SwapchainSupportDetails querySwapchainSupport(VkPhysicalDevice device, VkSurfaceKHR surface)
-{
-	SwapchainSupportDetails details = {};
-
-	vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &(details.capabilities));
-
-	unsigned int formatCount;
-	vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, nullptr);
-
-	if (formatCount > 0)
-	{
-		details.formats.resize(formatCount);
-		vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, details.formats.data());
-	}
-
-	unsigned int presentModeCount;
-	vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, nullptr);
-
-	if (presentModeCount > 0)
-	{
-		details.presentModes.resize(presentModeCount);
-		vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, details.presentModes.data());
-	}
-
-	return details;
-}
-
-VkSurfaceFormatKHR chooseSwapSurfaceFormat(const std::vector<VkSurfaceFormatKHR>& formats)
-{
-	for (const VkSurfaceFormatKHR& surfaceFormat : formats)
-	{
-		if (surfaceFormat.format == VK_FORMAT_B8G8R8A8_SRGB && surfaceFormat.colorSpace == VK_COLOR_SPACE_SRGB_NONLINEAR_KHR)
-		{
-			return surfaceFormat;
-		}
-	}
-
-	return formats[0];
-}
-
-VkPresentModeKHR choosePresentMode(const std::vector<VkPresentModeKHR>& modes, bool vsync)
-{
-	for (const VkPresentModeKHR& mode : modes)
-	{
-		// If vsync is disabled, prefer immediate.
-		if (vsync == false && mode == VK_PRESENT_MODE_IMMEDIATE_KHR)
-			return mode;
-
-		// If vsync is enabled, prefer triple buffering over double buffering.
-		if (vsync == true && mode == VK_PRESENT_MODE_MAILBOX_KHR)
-			return mode;
-	}
-
-	// If the preferred modes aren't available, default to double buffering.
-	return VK_PRESENT_MODE_FIFO_KHR;
-}
-
-VkExtent2D chooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities, const glacier::Window& window)
-{
-	if (capabilities.currentExtent.width != UINT32_MAX)
-	{
-		// If screen coordinates match pixels, return the current extent in screen coordinates.
-		return capabilities.currentExtent;
-	}
-	else
-	{
-		// If screen coordinates don't match pixels, get the current extent in pixels from GLFW
-		glm::uvec2 size = window.getFramebufferSize();
-
-		VkExtent2D actualExtent = { static_cast<uint32_t>(size.x), static_cast<uint32_t>(size.y) };
-
-		return actualExtent;
-	}
-}
-
 bool isDeviceSuitable(VkPhysicalDevice device, VkSurfaceKHR surface)
 {
 	VkPhysicalDeviceProperties deviceProperties;
@@ -229,480 +82,8 @@ bool isDeviceSuitable(VkPhysicalDevice device, VkSurfaceKHR surface)
 	return true;
 }
 
-/* Recreate swapchain */
-// Create swapchain
-void createSwapchain(const VkPhysicalDevice& physicalDevice, const VkDevice& device, const VkSurfaceKHR& surface, const glacier::ApplicationInfo& applicationInfo, const glacier::Window& window, const SwapchainSupportDetails& details, const VkSurfaceFormatKHR& surfaceFormat, const VkPresentModeKHR& presentMode, const VkExtent2D& extent, const VkSwapchainKHR* oldSwapchain, VkSwapchainKHR* pSwapchain)
-{
-	/* Create a swap chain */
-	glacier::g_Logger->trace("Creating swap chain");
-
-	unsigned int imageCount = details.capabilities.minImageCount + 1;
-	if (details.capabilities.minImageCount > 0 && imageCount > details.capabilities.maxImageCount)
-	{
-		imageCount = details.capabilities.maxImageCount;
-	}
-
-	VkSwapchainCreateInfoKHR swapchainCreateInfo = {};
-	swapchainCreateInfo.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
-	swapchainCreateInfo.surface = surface;
-	swapchainCreateInfo.minImageCount = imageCount;
-	swapchainCreateInfo.imageFormat = surfaceFormat.format;
-	swapchainCreateInfo.imageColorSpace = surfaceFormat.colorSpace;
-	swapchainCreateInfo.imageExtent = extent;
-	swapchainCreateInfo.imageArrayLayers = 1;
-
-	// Images will be written to directly
-	swapchainCreateInfo.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-
-	QueueFamilyIndices queueFamilyIndices = findQueueFamilies(physicalDevice, surface);
-
-	uint32_t indices[] = { queueFamilyIndices.graphicsFamily.value(), queueFamilyIndices.presentationFamily.value() };
-
-	if (queueFamilyIndices.graphicsFamily != queueFamilyIndices.presentationFamily)
-	{
-		// If the graphics and presentation families are different, we need to specify that both share ownership of the image.
-
-		swapchainCreateInfo.imageSharingMode = VK_SHARING_MODE_CONCURRENT;
-		swapchainCreateInfo.queueFamilyIndexCount = 2;
-		swapchainCreateInfo.pQueueFamilyIndices = indices;
-	}
-	else
-	{
-		// If the graphcis and presentation families are the same, we can safely specify that they exclusively own the image.
-		swapchainCreateInfo.imageSharingMode = VK_SHARING_MODE_EXCLUSIVE;
-		swapchainCreateInfo.queueFamilyIndexCount = 0;
-		swapchainCreateInfo.pQueueFamilyIndices = nullptr;
-	}
-
-	// Don't transform the image
-	swapchainCreateInfo.preTransform = details.capabilities.currentTransform;
-
-	// Ignore the alpha channel
-	swapchainCreateInfo.compositeAlpha = VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR;
-
-	swapchainCreateInfo.presentMode = presentMode;
-
-	// Enable clipping of unused pixels for better performance
-	swapchainCreateInfo.clipped = VK_TRUE;
-
-	// TODO: If the window is resized, we need to create an entirely new swapchain
-	swapchainCreateInfo.oldSwapchain = VK_NULL_HANDLE;
-
-	if (vkCreateSwapchainKHR(device, &swapchainCreateInfo, nullptr, pSwapchain) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create swapchain");
-	}
-}
-
-// Create image views
-void createImageViews(const VkDevice& device, const VkSurfaceFormatKHR& surfaceFormat, const VkExtent2D& extent, const VkSwapchainKHR& swapchain, std::vector<VkImage>& swapchainImages, std::vector<VkImageView>& imageViews)
-{
-	glacier::g_Logger->trace("Creating image views...");
-
-	VkFormat swapchainImageFormat = surfaceFormat.format;
-	VkExtent2D swapchainExtent = extent;
-
-	uint32_t imageCount;
-	vkGetSwapchainImagesKHR(device, swapchain, &imageCount, nullptr);
-
-	swapchainImages.clear();
-	swapchainImages.resize(imageCount);
-	vkGetSwapchainImagesKHR(device, swapchain, &imageCount, swapchainImages.data());
-
-	imageViews.clear();
-	imageViews.resize(swapchainImages.size());
-	for (size_t i = 0; i < swapchainImages.size(); i++)
-	{
-		VkImageViewCreateInfo imageViewCreateInfo = {};
-		imageViewCreateInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-		imageViewCreateInfo.image = swapchainImages[i];
-		imageViewCreateInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-		imageViewCreateInfo.format = swapchainImageFormat;
-		imageViewCreateInfo.components.r = VK_COMPONENT_SWIZZLE_IDENTITY;
-		imageViewCreateInfo.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-		imageViewCreateInfo.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-		imageViewCreateInfo.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-
-		imageViewCreateInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-		imageViewCreateInfo.subresourceRange.baseMipLevel = 0;
-		imageViewCreateInfo.subresourceRange.levelCount = 1;
-		imageViewCreateInfo.subresourceRange.baseArrayLayer = 0;
-		imageViewCreateInfo.subresourceRange.layerCount = 1;
-
-		if (vkCreateImageView(device, &imageViewCreateInfo, nullptr, &(*(imageViews.begin() + i))) != VK_SUCCESS)
-		{
-			throw std::runtime_error("Failed to create image view");
-		}
-	}
-}
-
-// Create render pass
-void createRenderPass(const VkDevice& device, const VkSurfaceFormatKHR& surfaceFormat, VkRenderPass* renderPass)
-{
-	glacier::g_Logger->trace("Creating render pass...");
-
-	// Create color attachment (To clear the screen)
-	VkAttachmentDescription colorAttachment = {};
-	colorAttachment.format = surfaceFormat.format;
-	colorAttachment.samples = VK_SAMPLE_COUNT_1_BIT;
-	colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-	colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-
-	// We won't do anything with the stencil buffer, so we ignore these
-	colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-	colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-
-	colorAttachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-	colorAttachment.finalLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-
-	// Create attachment reference
-	VkAttachmentReference colorAttachmentReference = {};
-	colorAttachmentReference.attachment = 0;
-	colorAttachmentReference.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-	// Create subpass
-	VkSubpassDescription subpassDescription = {};
-	subpassDescription.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-	subpassDescription.colorAttachmentCount = 1;
-	subpassDescription.pColorAttachments = &colorAttachmentReference;
-
-	// Create render pass
-	VkRenderPassCreateInfo renderPassCreateInfo = {};
-	renderPassCreateInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-	renderPassCreateInfo.attachmentCount = 1;
-	renderPassCreateInfo.pAttachments = &colorAttachment;
-	renderPassCreateInfo.subpassCount = 1;
-	renderPassCreateInfo.pSubpasses = &subpassDescription;
-
-	if (vkCreateRenderPass(device, &renderPassCreateInfo, nullptr, renderPass) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create render pass");
-	}
-}
-
-// Create graphics pipeline
-void createPipeline(const VkDevice& device, const glacier::Window& window, const VkRenderPass& renderPass, const BufferInfo& vertexBuffer, const std::vector<ShaderInfo>& shaders, VkPipeline* pipeline, VkPipelineLayout* pipelineLayout)
-{
-	glacier::g_Logger->trace("Creating pipeline...");
-
-	std::vector<VkPipelineShaderStageCreateInfo> shaderStages;
-
-	bool hasVertex = false, hasFragment = false;
-
-	for (const ShaderInfo& info : shaders)
-	{
-		VkPipelineShaderStageCreateInfo shaderCreateInfo = {};
-		shaderCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-		shaderCreateInfo.stage = info.stage;
-
-		if (info.stage & VK_SHADER_STAGE_VERTEX_BIT)
-		{
-			if (!hasVertex)
-				hasVertex = true;
-			else
-				throw std::runtime_error("You can't have more than one vertex shader");
-		}
-		
-		if (info.stage & VK_SHADER_STAGE_FRAGMENT_BIT)
-		{
-			if (!hasFragment)
-				hasFragment = true;
-			else
-				throw std::runtime_error("You can't have more than one fragment shader");
-		}
-
-		shaderCreateInfo.module = info.module;
-		shaderCreateInfo.pName = "main";
-
-		shaderStages.push_back(shaderCreateInfo);
-	}
-
-	if (!hasVertex || !hasFragment)
-		throw std::runtime_error("At least one vertex and fragment shader must exist");
-
-	VkPipelineInputAssemblyStateCreateInfo inputAssemblyCreateInfo = {};
-	inputAssemblyCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-	inputAssemblyCreateInfo.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-	inputAssemblyCreateInfo.primitiveRestartEnable = VK_FALSE;
-
-	VkPipelineVertexInputStateCreateInfo vertexInputCreateInfo = {};
-	vertexInputCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-
-	// TODO: Vertex buffers
-
-	// EDIT
-	//vertexInputCreateInfo.vertexBindingDescriptionCount = 0;
-	//vertexInputCreateInfo.pVertexBindingDescriptions = nullptr;
-	//vertexInputCreateInfo.vertexAttributeDescriptionCount = 0;
-	//vertexInputCreateInfo.pVertexAttributeDescriptions = nullptr;
-	// ->
-	
-	vertexInputCreateInfo.vertexBindingDescriptionCount = 1;
-	vertexInputCreateInfo.pVertexBindingDescriptions = &vertexBuffer.inputBinding;
-	vertexInputCreateInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(vertexBuffer.inputAttributes.size());
-	vertexInputCreateInfo.pVertexAttributeDescriptions = vertexBuffer.inputAttributes.data();
-
-	/*
-	 * float:	VK_FORMAT_R32_SFLOAT
-	 * vec2:	VK_FORMAT_R32G32_SFLOAT
-	 * vec3:	VK_FORMAT_R32G32B32_SFLOAT
-	 * vec4:	VK_FORMAT_R32G32B32A32_SFLOAT
-	 */
-	// END
-
-	VkViewport viewport = {};
-	viewport.x = 0.0f;
-	viewport.y = 0.0f;
-
-	glm::uvec2 size = window.getFramebufferSize();
-	viewport.width = static_cast<float>(size.x);
-	viewport.height = static_cast<float>(size.y);
-	viewport.minDepth = 0.0f;
-	viewport.maxDepth = 1.0f;
-
-	VkRect2D scissor = {};
-	scissor.offset = { 0, 0 };
-	scissor.extent = { size.x, size.y };
-
-	VkPipelineViewportStateCreateInfo viewportCreateInfo = {};
-	viewportCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-	viewportCreateInfo.viewportCount = 1;
-	viewportCreateInfo.pViewports = &viewport;
-	viewportCreateInfo.scissorCount = 1;
-	viewportCreateInfo.pScissors = &scissor;
-
-	VkPipelineRasterizationStateCreateInfo rasterizerCreateInfo = {};
-	rasterizerCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-	rasterizerCreateInfo.depthClampEnable = VK_FALSE;
-	rasterizerCreateInfo.rasterizerDiscardEnable = VK_FALSE;
-	rasterizerCreateInfo.polygonMode = VK_POLYGON_MODE_FILL;
-	rasterizerCreateInfo.lineWidth = 1.0f;
-
-	// Disable face culling
-	rasterizerCreateInfo.cullMode = VK_CULL_MODE_NONE; // VK_CULL_MODE_BACK_BIT
-
-	rasterizerCreateInfo.frontFace = VK_FRONT_FACE_CLOCKWISE;
-	rasterizerCreateInfo.depthBiasEnable = VK_FALSE;
-	rasterizerCreateInfo.depthBiasConstantFactor = 0.0f;
-	rasterizerCreateInfo.depthBiasClamp = 0.0f;
-	rasterizerCreateInfo.depthBiasSlopeFactor = 0.0f;
-
-	VkPipelineMultisampleStateCreateInfo multisampleCreateInfo = {};
-	multisampleCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-	multisampleCreateInfo.sampleShadingEnable = VK_FALSE;
-	multisampleCreateInfo.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-	multisampleCreateInfo.minSampleShading = 1.0f;
-	multisampleCreateInfo.pSampleMask = nullptr;
-	multisampleCreateInfo.alphaToCoverageEnable = VK_FALSE;
-	multisampleCreateInfo.alphaToOneEnable = VK_FALSE;
-
-	VkPipelineColorBlendAttachmentState colorBlendAttachment = {};
-	colorBlendAttachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-	colorBlendAttachment.blendEnable = VK_FALSE;
-	colorBlendAttachment.srcColorBlendFactor = VK_BLEND_FACTOR_ONE;
-	colorBlendAttachment.dstColorBlendFactor = VK_BLEND_FACTOR_ZERO;
-	colorBlendAttachment.colorBlendOp = VK_BLEND_OP_ADD;
-	colorBlendAttachment.srcAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-	colorBlendAttachment.dstAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-
-	VkPipelineColorBlendStateCreateInfo colorBlendCreateInfo = {};
-	colorBlendCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-	colorBlendCreateInfo.logicOpEnable = VK_FALSE;
-	colorBlendCreateInfo.logicOp = VK_LOGIC_OP_COPY;
-	colorBlendCreateInfo.attachmentCount = 1;
-	colorBlendCreateInfo.pAttachments = &colorBlendAttachment;
-	colorBlendCreateInfo.blendConstants[0] = 0.0f;
-	colorBlendCreateInfo.blendConstants[1] = 0.0f;
-	colorBlendCreateInfo.blendConstants[2] = 0.0f;
-	colorBlendCreateInfo.blendConstants[3] = 0.0f;
-
-	VkPipelineLayoutCreateInfo pipelineLayoutCreateInfo = {};
-	pipelineLayoutCreateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-	pipelineLayoutCreateInfo.setLayoutCount = 0;
-	pipelineLayoutCreateInfo.pSetLayouts = nullptr;
-	pipelineLayoutCreateInfo.pushConstantRangeCount = 0;
-	pipelineLayoutCreateInfo.pPushConstantRanges = nullptr;
-
-	if (vkCreatePipelineLayout(device, &pipelineLayoutCreateInfo, nullptr, pipelineLayout) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create pipeline layout");
-	}
-
-	/* Create graphics pipeline */
-	VkGraphicsPipelineCreateInfo graphicsPipelineCreateInfo = {};
-	graphicsPipelineCreateInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-
-	// Shader stages
-	graphicsPipelineCreateInfo.stageCount = static_cast<uint32_t>(shaderStages.size());
-	graphicsPipelineCreateInfo.pStages = shaderStages.data();
-
-	graphicsPipelineCreateInfo.pVertexInputState = &vertexInputCreateInfo;
-	graphicsPipelineCreateInfo.pInputAssemblyState = &inputAssemblyCreateInfo;
-	graphicsPipelineCreateInfo.pViewportState = &viewportCreateInfo;
-	graphicsPipelineCreateInfo.pRasterizationState = &rasterizerCreateInfo;
-	graphicsPipelineCreateInfo.pMultisampleState = &multisampleCreateInfo;
-	graphicsPipelineCreateInfo.pDepthStencilState = nullptr;
-	graphicsPipelineCreateInfo.pColorBlendState = &colorBlendCreateInfo;
-
-	// TODO: Implement dynamic state (To change viewport size)
-	graphicsPipelineCreateInfo.pDynamicState = nullptr;
-
-	graphicsPipelineCreateInfo.layout = *pipelineLayout;
-
-	graphicsPipelineCreateInfo.renderPass = renderPass;
-	graphicsPipelineCreateInfo.subpass = 0;
-
-	graphicsPipelineCreateInfo.basePipelineHandle = VK_NULL_HANDLE;
-	graphicsPipelineCreateInfo.basePipelineIndex = -1;
-
-	if (vkCreateGraphicsPipelines(device, VK_NULL_HANDLE, 1, &graphicsPipelineCreateInfo, nullptr, pipeline) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create graphics pipeline");
-	}
-}
-// Create framebuffers
-void createFramebuffers(const VkDevice& device, const std::vector<VkImageView>& imageViews, const VkRenderPass& renderPass, const VkExtent2D& swapchainExtent, std::vector<VkFramebuffer>& framebuffers)
-{
-	glacier::g_Logger->trace("Creating framebuffers...");
-
-	framebuffers.clear();
-	framebuffers.resize(imageViews.size());
-
-	// Create a framebuffer for every image view
-	for (size_t i = 0; i < imageViews.size(); i++)
-	{
-		std::vector<VkImageView> attachments = {
-			imageViews[i]
-		};
-
-		VkFramebufferCreateInfo framebufferCreateInfo = {};
-		framebufferCreateInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
-		framebufferCreateInfo.renderPass = renderPass;
-		framebufferCreateInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
-		framebufferCreateInfo.pAttachments = attachments.data();
-		framebufferCreateInfo.width = swapchainExtent.width;
-		framebufferCreateInfo.height = swapchainExtent.height;
-		framebufferCreateInfo.layers = 1;
-
-		if (vkCreateFramebuffer(device, &framebufferCreateInfo, nullptr, &framebuffers[i]) != VK_SUCCESS)
-		{
-			throw std::runtime_error(fmt::format("Failed to create framebuffer {}", i));
-		}
-	}
-}
-// Create command pool
-void createCommandPool(const VkDevice& device, const QueueFamilyIndices& queueFamilyIndices, VkCommandPool* commandPool)
-{
-	glacier::g_Logger->trace("Creating command pool");
-
-	VkCommandPoolCreateInfo commandPoolCreateInfo = {};
-	commandPoolCreateInfo.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-	commandPoolCreateInfo.queueFamilyIndex = queueFamilyIndices.graphicsFamily.value();
-	commandPoolCreateInfo.flags = 0;
-
-	if (vkCreateCommandPool(device, &commandPoolCreateInfo, nullptr, commandPool) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create command pool");
-	}
-}
-// Create command buffers
-void createCommandBuffers(const VkDevice& device, const std::vector<VkFramebuffer>& framebuffers, const VkCommandPool& commandPool, const VkRenderPass& renderPass, const VkExtent2D& extent, const VkPipeline& pipeline, const VkBuffer& vertexBuffer, uint32_t vertices, std::vector<VkCommandBuffer>& commandBuffers)
-{
-	glacier::g_Logger->trace("Creating command buffers...");
-
-	commandBuffers.clear();
-	commandBuffers.resize(framebuffers.size());
-
-	VkCommandBufferAllocateInfo commandBufferAllocateInfo = {};
-	commandBufferAllocateInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-	commandBufferAllocateInfo.commandPool = commandPool;
-	commandBufferAllocateInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-	commandBufferAllocateInfo.commandBufferCount = static_cast<uint32_t>(commandBuffers.size());
-
-	if (vkAllocateCommandBuffers(device, &commandBufferAllocateInfo, commandBuffers.data()) != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to allocate command buffers");
-	}
-
-	for (size_t i = 0; i < commandBuffers.size(); i++)
-	{
-		VkCommandBufferBeginInfo commandBufferBeginInfo = {};
-		commandBufferBeginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-		commandBufferBeginInfo.flags = 0;
-		commandBufferBeginInfo.pInheritanceInfo = nullptr;
-
-		if (vkBeginCommandBuffer(commandBuffers[i], &commandBufferBeginInfo) != VK_SUCCESS)
-		{
-			throw std::runtime_error("Failed to begin command buffer");
-		}
-
-		// Begin render pass
-		VkRenderPassBeginInfo renderPassBeginInfo = {};
-		renderPassBeginInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
-		renderPassBeginInfo.renderPass = renderPass;
-		renderPassBeginInfo.framebuffer = framebuffers[i];
-
-		renderPassBeginInfo.renderArea.offset = { 0, 0 };
-		renderPassBeginInfo.renderArea.extent = extent;
-
-		VkClearValue clearColor = { 0.0f, 0.0f, 0.0f, 1.0f };
-		renderPassBeginInfo.clearValueCount = 1;
-		renderPassBeginInfo.pClearValues = &clearColor;
-
-		vkCmdBeginRenderPass(commandBuffers[i], &renderPassBeginInfo, VK_SUBPASS_CONTENTS_INLINE);
-
-		vkCmdBindPipeline(commandBuffers[i], VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline);
-
-		// $ BEGIN $
-		VkBuffer vertexBuffers[] = { vertexBuffer };
-		VkDeviceSize offsets[] = { 0 };
-		vkCmdBindVertexBuffers(commandBuffers[i], 0, 1, vertexBuffers, offsets);
-		// $ END $
-
-		vkCmdDraw(commandBuffers[i], vertices, 1, 0, 0);
-
-		vkCmdEndRenderPass(commandBuffers[i]);
-
-		if (vkEndCommandBuffer(commandBuffers[i]) != VK_SUCCESS)
-		{
-			throw std::runtime_error("Failed to end command buffer");
-		}
-	}
-}
-// Destroy swapchain
-void destroySwapchain(const VkDevice& device, std::vector<VkFramebuffer>& framebuffers, const VkCommandPool& commandPool, std::vector<VkCommandBuffer>& commandBuffers, VkPipeline* pipeline, VkPipelineLayout* pipelineLayout, VkRenderPass* renderPass, std::vector<VkImageView>& imageViews, VkSwapchainKHR* swapchain)
-{
-	for (size_t i = 0; i < framebuffers.size(); i++)
-	{
-		vkDestroyFramebuffer(device, framebuffers[i], nullptr);
-	}
-
-	vkFreeCommandBuffers(device, commandPool, static_cast<uint32_t>(commandBuffers.size()), commandBuffers.data());
-	commandBuffers.clear();
-
-	vkDestroyPipeline(device, *pipeline, nullptr);
-	*pipeline = nullptr;
-
-	vkDestroyPipelineLayout(device, *pipelineLayout, nullptr);
-	*pipelineLayout = nullptr;
-
-	vkDestroyRenderPass(device, *renderPass, nullptr);
-	*renderPass = nullptr;
-
-	for (size_t i = 0; i < imageViews.size(); i++)
-	{
-		vkDestroyImageView(device, imageViews[i], nullptr);
-	}
-
-	imageViews.clear();
-
-	vkDestroySwapchainKHR(device, *swapchain, nullptr);
-	*swapchain = nullptr;
-}
-/* ------------------ */
-
 glacier::Application::Application(const ApplicationInfo& info)
-	: m_Info(info)
+	: m_Info(info), m_FramebufferResized(false), m_Renderer(nullptr)
 {
 	g_Logger->info("Initializing application...");
 
@@ -773,7 +154,7 @@ glacier::Application::Application(const ApplicationInfo& info)
 
 	/* Create a window surface */
 	g_Logger->debug("Creating window surface...");
-	if (glfwCreateWindowSurface(static_cast<VkInstance>(m_VulkanInstance), static_cast<GLFWwindow*>(m_Window->m_Window), nullptr, reinterpret_cast<VkSurfaceKHR*>(&m_Surface)) != VK_SUCCESS)
+	if (glfwCreateWindowSurface(static_cast<VkInstance>(m_VulkanInstance), static_cast<GLFWwindow*>(m_Window->m_Handle), nullptr, reinterpret_cast<VkSurfaceKHR*>(&m_Surface)) != VK_SUCCESS)
 	{
 		throw std::runtime_error("Failed to create window surface");
 	}
@@ -877,88 +258,16 @@ glacier::Application::~Application()
 
 void glacier::Application::run()
 {
-	g_Logger->debug("Creating swapchain...");
+	initialize();
 
-	/* Get queue family indices */
-	QueueFamilyIndices queueFamilyIndices = findQueueFamilies(static_cast<VkPhysicalDevice>(m_PhysicalDevice), static_cast<VkSurfaceKHR>(m_Surface));
-
-	/* Create swapchain */
-	SwapchainSupportDetails details = querySwapchainSupport(static_cast<VkPhysicalDevice>(m_PhysicalDevice), static_cast<VkSurfaceKHR>(m_Surface));
-
-	VkSurfaceFormatKHR surfaceFormat = chooseSwapSurfaceFormat(details.formats);
-	VkPresentModeKHR presentMode = choosePresentMode(details.presentModes, m_Info.vsync);
-	VkExtent2D extent = chooseSwapExtent(details.capabilities, *m_Window);
-
-	VkSwapchainKHR swapchain;
-	createSwapchain(static_cast<VkPhysicalDevice>(m_PhysicalDevice), static_cast<VkDevice>(m_Device), static_cast<VkSurfaceKHR>(m_Surface), m_Info, *m_Window, details, surfaceFormat, presentMode, extent, nullptr, &swapchain);
-
-	/* Create image views */
-	std::vector<VkImage> swapchainImages;
-	std::vector<VkImageView> imageViews;
-	createImageViews(static_cast<VkDevice>(m_Device), surfaceFormat, extent, swapchain, swapchainImages, imageViews);
-
-	/* Create render pass */
-	VkRenderPass renderPass;
-	createRenderPass(static_cast<VkDevice>(m_Device), surfaceFormat, &renderPass);
-
-	/* Create pipeline */
-	glacier::PipelineInfo pipelineInfo = {};
-	this->initialize(pipelineInfo);
-
-	std::vector<ShaderInfo> shaderInfos;
-	std::transform(pipelineInfo.shaders.begin(), pipelineInfo.shaders.end(), std::back_inserter(shaderInfos), [](Shader* shader)
-		{
-			ShaderInfo info = {};
-			info.module = static_cast<VkShaderModule>((*shader).m_ShaderModule);
-
-			switch (shader->getType())
-			{
-			case ShaderType::Vertex:
-				info.stage = VK_SHADER_STAGE_VERTEX_BIT;
-				break;
-			case ShaderType::Fragment:
-				info.stage = VK_SHADER_STAGE_FRAGMENT_BIT;
-				break;
-			}
-
-			return info;
-		});
-
-	VkPipeline pipeline;
-	VkPipelineLayout pipelineLayout;
-	
-	if (pipelineInfo.vertexBuffer == nullptr)
-		throw std::runtime_error("No vertex buffer provided in pipeline info");
-
-	VkVertexInputBindingDescription vertexInputBindingDescription = pipelineInfo.vertexBuffer->m_Layout.getBindingDescription();
-	std::vector<VkVertexInputAttributeDescription> vertexInputAttributeDescriptions = pipelineInfo.vertexBuffer->m_Layout.getAttributeDescriptions();
-
-	BufferInfo vertexBuffer = {
-		pipelineInfo.vertexBuffer->m_Layout.getBindingDescription(),
-		pipelineInfo.vertexBuffer->m_Layout.getAttributeDescriptions()
-	};
-
-	createPipeline(static_cast<VkDevice>(m_Device), *m_Window, renderPass, vertexBuffer, shaderInfos, &pipeline, &pipelineLayout);
-
-	/* Create framebuffers */
-	std::vector<VkFramebuffer> framebuffers;
-	createFramebuffers(static_cast<VkDevice>(m_Device), imageViews, renderPass, extent, framebuffers);
-
-	/* Create command pool */
-	VkCommandPool commandPool;
-	createCommandPool(static_cast<VkDevice>(m_Device), queueFamilyIndices, &commandPool);
-
-	/* Create command buffers */
-	std::vector<VkCommandBuffer> commandBuffers;
-	createCommandBuffers(static_cast<VkDevice>(m_Device), framebuffers, commandPool, renderPass, extent, pipeline, static_cast<VkBuffer>(pipelineInfo.vertexBuffer->m_Handle), pipelineInfo.vertexCount, commandBuffers);
-
-	/* RENDER */
+	m_Renderer = new Renderer(this);
+	initializeRenderer(m_Renderer);
 
 	/* Create semaphores */
 	std::vector<VkSemaphore> imageAvailableSemaphores(MAX_BUFFERED_FRAMES);
 	std::vector<VkSemaphore> renderFinishedSemaphores(MAX_BUFFERED_FRAMES);
 	std::vector<VkFence> bufferedFences(MAX_BUFFERED_FRAMES);
-	std::vector<VkFence> bufferedImageFences(swapchainImages.size(), VK_NULL_HANDLE);
+	std::vector<VkFence> bufferedImageFences(m_Renderer->m_Images.size(), VK_NULL_HANDLE);
 
 	VkSemaphoreCreateInfo semaphoreCreateInfo = {};
 	semaphoreCreateInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
@@ -973,7 +282,7 @@ void glacier::Application::run()
 		{
 			throw std::runtime_error("Failed to create image available semaphore");
 		}
-		
+
 		if (vkCreateSemaphore(static_cast<VkDevice>(m_Device), &semaphoreCreateInfo, nullptr, &(renderFinishedSemaphores[i])) != VK_SUCCESS)
 		{
 			throw std::runtime_error("Failed to create render finished semaphore");
@@ -985,36 +294,50 @@ void glacier::Application::run()
 		}
 	}
 
-	/* Get queues */
-	VkQueue graphicsQueue;
-	vkGetDeviceQueue(static_cast<VkDevice>(m_Device), queueFamilyIndices.graphicsFamily.value(), 0, &graphicsQueue);
-
-	VkQueue presentQueue;
-	vkGetDeviceQueue(static_cast<VkDevice>(m_Device), queueFamilyIndices.presentationFamily.value(), 0, &presentQueue);
-
 	/* Start game loop */
 	g_Logger->debug("Starting game loop...");
 
 	size_t currentFrame = 0;
+	bool suboptimal_flag = false;
+
+	glfwSetWindowUserPointer(static_cast<GLFWwindow*>(m_Window->m_Handle), &m_FramebufferResized);
+	glfwSetFramebufferSizeCallback(static_cast<GLFWwindow*>(m_Window->m_Handle), [](GLFWwindow* window, int width, int height) -> void
+		{
+			bool* framebufferResized = reinterpret_cast<bool*>(glfwGetWindowUserPointer(window));
+			*framebufferResized = true;
+		});
 
 	double lastTime = glfwGetTime();
 	while (m_Window->isOpen())
 	{
-		// Wait until the next frame should be drawn
-		vkWaitForFences(static_cast<VkDevice>(m_Device), 1, &(bufferedFences[currentFrame]), VK_TRUE, UINT64_MAX);
-
 		double deltaTime = glfwGetTime() - lastTime;
 		lastTime = glfwGetTime();
 
 		update(deltaTime);
-		render();
 
 		/* Draw frame */
-		uint32_t imageIndex;
-		VkResult result = vkAcquireNextImageKHR(static_cast<VkDevice>(m_Device), swapchain, UINT64_MAX, imageAvailableSemaphores[currentFrame], VK_NULL_HANDLE, &imageIndex);
+		// Wait until the next frame should be drawn
+		vkWaitForFences(static_cast<VkDevice>(m_Device), 1, &(bufferedFences[currentFrame]), VK_TRUE, UINT64_MAX);
 
-		if (result == VK_ERROR_OUT_OF_DATE_KHR)
+		uint32_t imageIndex;
+		VkResult result = vkAcquireNextImageKHR(static_cast<VkDevice>(m_Device), static_cast<VkSwapchainKHR>(m_Renderer->m_Swapchain), UINT64_MAX, imageAvailableSemaphores[currentFrame], VK_NULL_HANDLE, &imageIndex);
+
+		if (result == VK_SUCCESS)
 		{
+			suboptimal_flag = false;
+		}
+		else if (result == VK_SUBOPTIMAL_KHR)
+		{
+			if (!suboptimal_flag)
+			{
+				g_Logger->warn("Swapchain is suboptimal.");
+				suboptimal_flag = true;
+			}
+		}
+		else if (result == VK_ERROR_OUT_OF_DATE_KHR || m_FramebufferResized)
+		{
+			m_FramebufferResized = false;
+
 			/* Check if the window was minimized */
 			glm::uvec2 size = m_Window->getFramebufferSize();
 			if (size.x == 0 || size.y == 0)
@@ -1031,34 +354,34 @@ void glacier::Application::run()
 			}
 
 			/* Recreate the swapchain */
-			glacier::g_Logger->debug("Recreating swapchain...");
+			glacier::g_Logger->trace("Swapchain is outdated.");
 
-			vkDeviceWaitIdle(static_cast<VkDevice>(m_Device));
+			terminateRenderer(m_Renderer);
 
-			// Destroy the current swapchain
-			destroySwapchain(static_cast<VkDevice>(m_Device), framebuffers, commandPool, commandBuffers, &pipeline, &pipelineLayout, &renderPass, imageViews, &swapchain);
+			delete m_Renderer;
+			m_Renderer = new Renderer(this);
 
-			details = querySwapchainSupport(static_cast<VkPhysicalDevice>(m_PhysicalDevice), static_cast<VkSurfaceKHR>(m_Surface));
+			initializeRenderer(m_Renderer);
 
-			surfaceFormat = chooseSwapSurfaceFormat(details.formats);
-			presentMode = choosePresentMode(details.presentModes, m_Info.vsync);
-			extent = chooseSwapExtent(details.capabilities, *m_Window);
+			/* Recreate semaphores */
+			for (size_t i = 0; i < MAX_BUFFERED_FRAMES; i++)
+			{
+				vkDestroySemaphore(static_cast<VkDevice>(m_Device), imageAvailableSemaphores[i], nullptr);
+				if (vkCreateSemaphore(static_cast<VkDevice>(m_Device), &semaphoreCreateInfo, nullptr, &(imageAvailableSemaphores[i])) != VK_SUCCESS)
+				{
+					throw std::runtime_error("Failed to create image available semaphore");
+				}
 
-			createSwapchain(static_cast<VkPhysicalDevice>(m_PhysicalDevice), static_cast<VkDevice>(m_Device), static_cast<VkSurfaceKHR>(m_Surface), m_Info, *m_Window, details, surfaceFormat, presentMode, extent, nullptr, &swapchain);
-
-			createImageViews(static_cast<VkDevice>(m_Device), surfaceFormat, extent, swapchain, swapchainImages, imageViews);
-
-			createRenderPass(static_cast<VkDevice>(m_Device), surfaceFormat, &renderPass);
-
-			createPipeline(static_cast<VkDevice>(m_Device), *m_Window, renderPass, vertexBuffer, shaderInfos, &pipeline, &pipelineLayout);
-
-			createFramebuffers(static_cast<VkDevice>(m_Device), imageViews, renderPass, extent, framebuffers);
-
-			createCommandBuffers(static_cast<VkDevice>(m_Device), framebuffers, commandPool, renderPass, extent, pipeline, static_cast<VkBuffer>(pipelineInfo.vertexBuffer->m_Handle), pipelineInfo.vertexCount, commandBuffers);
+				vkDestroySemaphore(static_cast<VkDevice>(m_Device), renderFinishedSemaphores[i], nullptr);
+				if (vkCreateSemaphore(static_cast<VkDevice>(m_Device), &semaphoreCreateInfo, nullptr, &(renderFinishedSemaphores[i])) != VK_SUCCESS)
+				{
+					throw std::runtime_error("Failed to create render finished semaphore");
+				}
+			}
 
 			continue;
 		}
-		else if (result != VK_SUCCESS && result != VK_SUBOPTIMAL_KHR)
+		else if (result != VK_SUCCESS)
 		{
 			throw std::runtime_error("Failed to acquire swapchain image");
 		}
@@ -1070,6 +393,9 @@ void glacier::Application::run()
 
 		bufferedImageFences[imageIndex] = bufferedFences[currentFrame];
 
+		// Render the frame
+		render(m_Renderer);
+
 		VkSubmitInfo submitInfo = {};
 		submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
 
@@ -1079,15 +405,19 @@ void glacier::Application::run()
 		submitInfo.pWaitSemaphores = waitSemaphores;
 		submitInfo.pWaitDstStageMask = waitStages;
 
+		// Renderer needs to have a pipeline bound
+		if (m_Renderer->m_CommandBuffers.empty())
+			throw std::runtime_error("Renderer has no pipeline bound");
+
 		submitInfo.commandBufferCount = 1;
-		submitInfo.pCommandBuffers = &commandBuffers[imageIndex];
+		submitInfo.pCommandBuffers = reinterpret_cast<VkCommandBuffer*>(&m_Renderer->m_CommandBuffers[imageIndex]); //&commandBuffers[imageIndex];
 
 		VkSemaphore signalSemaphores[] = { renderFinishedSemaphores[currentFrame] };
 		submitInfo.signalSemaphoreCount = 1;
 		submitInfo.pSignalSemaphores = signalSemaphores;
 
 		vkResetFences(static_cast<VkDevice>(m_Device), 1, &(bufferedFences[currentFrame]));
-		if (vkQueueSubmit(graphicsQueue, 1, &submitInfo, bufferedFences[currentFrame]) != VK_SUCCESS)
+		if (vkQueueSubmit(static_cast<VkQueue>(m_Renderer->m_GraphicsQueue), 1, &submitInfo, bufferedFences[currentFrame]) != VK_SUCCESS)
 		{
 			throw std::runtime_error("Failed to submit draw command buffer");
 		}
@@ -1097,18 +427,14 @@ void glacier::Application::run()
 		presentInfo.waitSemaphoreCount = 1;
 		presentInfo.pWaitSemaphores = signalSemaphores;
 
-		VkSwapchainKHR swapchains[] = { swapchain };
+		VkSwapchainKHR swapchains[] = { static_cast<VkSwapchainKHR>(m_Renderer->m_Swapchain) };
 		presentInfo.swapchainCount = 1;
 		presentInfo.pSwapchains = swapchains; // TODO: &swapchain ?
 		presentInfo.pImageIndices = &imageIndex;
 
-		vkQueuePresentKHR(presentQueue, &presentInfo);
+		vkQueuePresentKHR(static_cast<VkQueue>(m_Renderer->m_PresentationQueue), &presentInfo);
 
 		currentFrame = (currentFrame + 1) % MAX_BUFFERED_FRAMES;
-
-		//vkDeviceWaitIdle(static_cast<VkDevice>(m_Device));
-
-		// https://vulkan-tutorial.com/en/Drawing_a_triangle/Drawing/Rendering_and_presentation#page_Subpass-dependencies
 
 		glfwPollEvents();
 	}
@@ -1121,24 +447,6 @@ void glacier::Application::run()
 	/* Destroy renderer */
 	g_Logger->debug("Destroying renderer...");
 
-	destroySwapchain(static_cast<VkDevice>(m_Device), framebuffers, commandPool, commandBuffers, &pipeline, &pipelineLayout, &renderPass, imageViews, &swapchain);
-
-	//for (const VkFramebuffer& framebuffer : framebuffers)
-	//{
-	//	vkDestroyFramebuffer(static_cast<VkDevice>(m_Device), framebuffer, nullptr);
-	//}
-	//
-	//vkDestroyPipeline(static_cast<VkDevice>(m_Device), pipeline, nullptr);
-	//vkDestroyPipelineLayout(static_cast<VkDevice>(m_Device), pipelineLayout, nullptr);
-	//vkDestroyRenderPass(static_cast<VkDevice>(m_Device), renderPass, nullptr);
-	//
-	//for (const VkImageView& imageView : imageViews)
-	//{
-	//	vkDestroyImageView(static_cast<VkDevice>(m_Device), imageView, nullptr);
-	//}
-	//
-	//vkDestroySwapchainKHR(static_cast<VkDevice>(m_Device), swapchain, nullptr);
-
 	// Destroy this shit last
 	for (size_t i = 0; i < MAX_BUFFERED_FRAMES; i++)
 	{
@@ -1147,7 +455,10 @@ void glacier::Application::run()
 		vkDestroyFence(static_cast<VkDevice>(m_Device), bufferedFences[i], nullptr);
 	}
 
-	vkDestroyCommandPool(static_cast<VkDevice>(m_Device), commandPool, nullptr);
+	terminateRenderer(m_Renderer);
+	delete m_Renderer;
+
+	terminate();
 }
 
 void glacier::Application::stop()
