+++ todo.txt
@@ -1,0 +1,3 @@
+TODO:
+	In glacier::Renderer::bindPipeline(const Pipeline&, uint32_t)
+	- Move vertices to glacier::Pipeline
